moved the get carer by patient ID models file to carer.js models file, updated port to env variable
