+++ routes/carers.js
@@ -1,35 +1,42 @@
 import { Router } from "express";
-import { getAllCarers, getCarerByID } from "../models/carers.js";
-import { getPatientsByCarerID } from "../models/patients.js";
+import {
+  getAllCarers,
+  getCarerByID,
+  getPatientsByCarerID,
+} from "../models/carers.js";
 import { validateAccessToken } from "../auth0-middleware.js";
 
 const router = Router();
 
 // GET all carers
 router.get("/", validateAccessToken, async function (req, res) {
-	const carersArr = await getAllCarers();
-	res.json({
-		success: true,
-		payload: carersArr,
-	});
+  const carersArr = await getAllCarers();
+  res.json({
+    success: true,
+    payload: carersArr,
+  });
 });
 
 // GET carer's info by carer_id
 router.get("/:carer_id", validateAccessToken, async function (req, res) {
-	const carer = await getCarerByID(req.params.carer_id);
-	res.json({
-		success: true,
-		payload: carer,
-	});
+  const carer = await getCarerByID(req.params.carer_id);
+  res.json({
+    success: true,
+    payload: carer,
+  });
 });
 
 // GET array of assigned patients
-router.get("/:carer_id/patients", validateAccessToken, async function (req, res) {
-	const patientsArr = await getPatientsByCarerID(req.params.carer_id);
-	res.json({
-		success: true,
-		payload: patientsArr,
-	});
-});
+router.get(
+  "/:carer_id/patients",
+  validateAccessToken,
+  async function (req, res) {
+    const patientsArr = await getPatientsByCarerID(req.params.carer_id);
+    res.json({
+      success: true,
+      payload: patientsArr,
+    });
+  }
+);
 
 export default router;
